move metrics to pkg
