+++ internal/remote/blob/azure_blob_storage.go
@@ -7,7 +7,7 @@ import (
 
 	"github.com/Azure/azure-sdk-for-go/sdk/storage/azblob/blob"
 	"github.com/Azure/azure-sdk-for-go/sdk/storage/azblob/blockblob"
-	"github.com/mazrean/gocica/internal/metrics"
+	"github.com/mazrean/gocica/internal/pkg/metrics"
 )
 
 var _ UploadClient = (*AzureUploadClient)(nil)
